rename session actions
