+++ frontend/components/auth/AuthContainer.js
@@ -1,5 +1,5 @@
 import { connect } from "react-redux";
-import { logout } from "../../actions/session_actions";
+import { logout } from "../../actions/sessionActions";
 import { openModal } from "../../actions/modalActions";
 import Auth from "./Auth";
