+++ frontend/components/login/login_form_container.jsx
@@ -2,7 +2,7 @@
 import React from "react";
 import { connect } from "react-redux";
 //remember to import the action you want to pass into props
-import { login, clearSessionErrors } from "../../actions/session_actions";
+import { login, clearSessionErrors } from "../../actions/sessionActions";
 //remember to import the presentational component itself
 import LoginForm from "./login_form";
 import { openModal, closeModal } from "../../actions/modalActions";
